feat(unicode): u32 cmap for astral plane emoji, Arabic + multilingual fonts
- Extend TTF cmap from HashMap<u16, u16> to HashMap<u32, u16> to support
  astral plane characters (emoji U+1F600+, CJK extensions, etc.)
- Parse cmap format 12 (segmented full Unicode) in addition to format 4
- Bundle NotoSans-Regular.ttf (Hebrew, Greek, Cyrillic) and
  NotoSansArabic-Regular.ttf for Arabic script rendering
- Fallback font chain: Arabic → Multilingual → Emoji → System CJK
- Split mixed-script runs at font-coverage boundaries
- All emoji render as monochrome outlines (Noto Emoji)

Known: fallback fonts are embedded without subsetting, causing large PDFs
when system fonts like ArialUnicodeMS are used. Next step: implement
proper per-glyph subsetting for fallback fonts.

+++ src/render/pdf.rs
@@ -8904,8 +8904,8 @@ mod tests {
         // Covers lines 1628-1657: TTF font objects in PDF
         use crate::parser::ttf::TtfFont;
         let mut cmap = HashMap::new();
-        for c in 32u16..=126 {
-            cmap.insert(c, c - 31);
+        for c in 32u32..=126 {
+            cmap.insert(c, (c - 31) as u16);
         }
         let ttf = TtfFont {
             font_name: "TestFont".to_string(),
@@ -8966,8 +8966,8 @@ mod tests {
         use crate::parser::ttf::TtfFont;
 
         let mut cmap = HashMap::new();
-        for c in 32u16..=126 {
-            cmap.insert(c, c - 31);
+        for c in 32u32..=126 {
+            cmap.insert(c, (c - 31) as u16);
         }
         let ttf = TtfFont {
             font_name: "TestFont".to_string(),
